+++ src/main/java/com/federation/federationdecollectivitesagricoles/entity/Position.java
@@ -1,32 +1,18 @@
 package com.federation.federationdecollectivitesagricoles.entity;
 
 public class Position {
-
-    private Integer id;
+    private Long id;
     private String name;
 
-    public Position() {
-    }
+    public Position() {}
 
-    public Position(Integer id, String name) {
+    public Position(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
+    public String getName() { return name; }
+    public void setName(String name) { this.name = name; }
 }
